backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-multipart
python-dotenv
tiktoken
langchain-text-splitters
langchain-core
beautifulsoup4
python-docx
docx2txt
